backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
cloudinary==1.40.0
pymongo==4.6.0
motor==3.3.2
openai==1.3.5
pydantic==2.5.0
requests==2.31.0
python-multipart==0.0.6
pillow==10.1.0
replicate==0.22.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
